+++ CHANGELOG.md
@@ -9,8 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Mostly automatic support `Ecto.Type` (including for builtin interval types)
+- `left/1` and `right/1` to extract the left and right values from the interval.
+- Builtin `Jason.Encoder` support.
+- `Interval.__using__` option `jason_encoder` for including encoder. Defaults to `true`.
 
 ### Changed
+
+- `Interval.__using__` option `discrete` is now optional, and defaults to `false`
 
 ### Deprecated