+++ CHANGELOG.md
@@ -8,6 +8,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+## [v0.3.0] - 2022-11-21
+
+### Added
+
 - Mostly automatic support `Ecto.Type` (including for builtin interval types)
 - `left/1` and `right/1` to extract the left and right values from the interval.
 - Builtin `Jason.Encoder` support.